// src/ibmmq/interface/ibmmq.interface.ts
export interface IIbmmqCreate {
  userId?: string;
  password?: string;
  connectionName?: string;
  channelName?: string;
  applName?: string;
  qManager?: string;
}
export interface IProducerConnection {
  userId?: string;
  password?: string;
  connectionName?: string;
  channelName?: string;
  applName?: string;
  qManager?: string;
  qName?: string;
}